refactor: move SidebarSize to constants
