+++ public/app/features/dashboard-scene/panel-edit/PanelEditNext/hooks.ts
@@ -11,7 +11,7 @@ import { PanelEditor } from '../PanelEditor';
 import { useSnappingSplitter } from '../splitter/useSnappingSplitter';
 import { useScrollReflowLimit } from '../useScrollReflowLimit';
 
-import { SidebarSize } from './QueryEditor/Sidebar/QueryEditorSidebar';
+import { SidebarSize } from './constants';
 
 type UseHorizontalResizeOptions = {
   initialWidth: number;
